MCR-3497 Move transformer classes to separate package
