+++ mycore-xeditor/src/main/java/org/mycore/frontend/xeditor/MCRXEditorTransformer.java
@@ -31,6 +31,7 @@ import org.mycore.common.content.transformer.MCRContentTransformerFactory;
 import org.mycore.common.content.transformer.MCRParameterizedTransformer;
 import org.mycore.common.content.transformer.MCRXSLTransformer;
 import org.mycore.common.xsl.MCRParameterCollector;
+import org.mycore.frontend.xeditor.transformer.MCRTransformerHelper;
 
 /**
  * @author Frank Lützenkirchen
