+++ mycore-xeditor/src/main/resources/components/xeditor/config/mycore.properties
@@ -45,7 +45,7 @@ MCR.ContentTransformer.xeditor.TransformerFactoryClass=%XALAN%
 MCR.ContentTransformer.xeditor.Stylesheet=xsl/xeditor-templates.xsl,xsl/xeditor.xsl
 
 # Transformer helper resolver to link between Java and XSL transformation of *.xed
-MCR.URIResolver.ModuleResolver.xedTransformerHelper=org.mycore.frontend.xeditor.MCRTransformerHelperResolver
+MCR.URIResolver.ModuleResolver.xedTransformerHelper=org.mycore.frontend.xeditor.transformer.MCRTransformerHelperResolver
 
 # Markers for input validation states. To be set in attributes, e.g. <div class="someclass {$xed-validation-marker}" ...
 # This marker is set if $xed-validation-marker is used in editor source code, but no validation was performed:
